Support non string path config property values
Previously the JSON parser would only accept strings as possible values for `properties`.
Because of this it was impossible to create a custom data structure inside the path-configuration file.
The iOS configuration parser does support non string values, matching parity with the other platform.

In doing so it allows for custom extension functions to be added to `PathConfigurationProperties` which allows
developers to serialize their own data structures.